requirements.txt
python-dotenv
gino
aiogram~=2.17
psycopg
SqlAlchemy~=1.2.16
